\name{MiRSEA-internal}
\alias{initializeMiRSEA}
\alias{Getenvir}
\title{MiRSEA internal functions}
\description{
  These functions internally called by other functions in the package \code{MiRSEA} and not meant to be called by the user.
}

\usage{

    initializeMiRSEA()
    Getenvir(envData)
}

\author{Junwei Han\email{hanjunwei1981@163.com},Siyao Liu \email{liusiyao29@163.com}}
\keyword{internal}